projects/web-crawler/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==22.1.0
aws-cdk-constructs==2.31.3
aws-cdk-lib==2.123.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-cdk.aws-neptune-alpha==2.123.0a0
AWSIoTPythonSDK==1.5.4
awslambdaric==2.0.11
b-aws-websocket-api==2.0.0
b-stage-deployment==1.0.1
beautifulsoup4==4.12.3
boto3==1.26.59
botocore==1.29.165
bs4==0.0.2
cattrs==23.1.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.3
constructs==10.3.0
coverage==7.2.3
dataclasses-json==0.6.7
diagrams==0.23.4
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
inquirerpy==0.3.4
Jinja2==3.1.4
jmespath==1.0.1
jsii==1.101.0
jsonpatch==1.33
jsonpointer==3.0.0
lambda-forge==2.1.10
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.36.1
orjson==3.10.6
packaging==24.1
pfzy==0.3.4
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
prompt_toolkit==3.0.47
publication==0.0.3
py==1.11.0
pydantic==2.7.3
pydantic_core==2.18.4
pyfiglet==1.0.2
pytest==6.2.5
pytest-cov==2.12.1
pytest-sugar==1.0.0
pytest-timeout==1.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
s3transfer==0.6.2
simplejson==3.18.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
tiktoken==0.7.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
typed-ast==1.5.5
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
zipp==3.19.2
